+++ java-client/src/test/java/org/opensearch/client/opensearch/core/SearchRequestTest.java
@@ -1,9 +1,21 @@
 package org.opensearch.client.opensearch.core;
 
+import java.util.Collections;
 import org.junit.Test;
+import org.opensearch.client.json.JsonData;
 import org.opensearch.client.opensearch.model.ModelTestCase;
 
 public class SearchRequestTest extends ModelTestCase {
+
+    @Test
+    public void ext() {
+        SearchRequest request = new SearchRequest.Builder().ext(
+            "similarity",
+            JsonData.of(Collections.singletonMap("fields", Collections.singletonList("name")))
+        ).build();
+
+        assertEquals("{\"ext\":{\"similarity\":{\"fields\":[\"name\"]}}}", toJson(request));
+    }
 
     @Test
     public void toBuilder() {